post 分页条件查询更改

+++ dingding-server/src/main/java/com/heima/dingding/service/impl/OrdersServiceImpl.java
@@ -7,8 +7,10 @@ import com.heima.dingdign.pojo.entity.BookCart;
 import com.heima.dingdign.pojo.entity.OrderDetail;
 import com.heima.dingdign.pojo.entity.Orders;
 import com.heima.dingdign.pojo.vo.OrderDetailVO;
+import com.heima.dingding.constant.MessageConstant;
 import com.heima.dingding.constant.PayConstant;
 import com.heima.dingding.context.BaseContext;
+import com.heima.dingding.exception.DataException;
 import com.heima.dingding.mapper.OrdersMapper;
 import com.heima.dingding.service.IBookCartService;
 import com.heima.dingding.service.IBookService;
@@ -38,6 +40,7 @@ public class OrdersServiceImpl extends ServiceImpl<OrdersMapper, Orders> impleme
 
     private final IBookCartService cartService;
     private final IBookService bookService;
+    private final IOrdersService orderService;
     private final IOrderDetailService detailService;
 
 
@@ -53,15 +56,21 @@ public class OrdersServiceImpl extends ServiceImpl<OrdersMapper, Orders> impleme
         List<BookCart> carts = cartService.lambdaQuery().eq(BookCart::getUserId, BaseContext.getCurrentId())
                 .list();
         //清除购物车
+        if (carts.size() <= 0) {
+            throw new DataException(MessageConstant.SHOPPING_CART_IS_NULL);
+        }
         cartService.removeBatchByIds(carts);
         //生成订单
         Orders orders = new Orders().setCreateTime(LocalDateTime.now())
                 .setUserId(BaseContext.getCurrentId())
                 .setStatus(PayConstant.NO_PAY);
         //保存
-        saveOrUpdate(orders);
+        save(orders);
         //查询订单是否生成成功
         orders = getById(orders);
+        if (orders == null) {
+            throw new DataException(MessageConstant.ORDER_NOT_FOUND);
+        }
         //添加订单详细数据
         List<OrderDetail> details = new ArrayList<>();
         for (BookCart bookCart : carts) {
@@ -71,13 +80,18 @@ public class OrdersServiceImpl extends ServiceImpl<OrdersMapper, Orders> impleme
             detail.setNumber(bookCart.getNumber());
             detail.setCreateTime(LocalDateTime.now());
             detail.setUpdateTime(LocalDateTime.now());
+            details.add(detail);
         }
-        detailService.saveBatch(details);
+        if (!detailService.saveBatch(details)) {
+            throw new DataException(MessageConstant.UPDATE_FAILED);
+        }
         return orders;
     }
 
     @Override
     public List<OrderDetailVO> getOrderDetail(Long orderId) {
+        //只有当前用户才可以看自己的订单
+
         //查询订单详情
         List<OrderDetail> details = detailService.lambdaQuery().eq(OrderDetail::getOrderId, orderId).list();
         //查询书籍信息
